+++ src/cmd/mod.rs
@@ -3,3 +3,4 @@ pub mod download;
 pub mod search;
 pub mod install;
 pub mod upself;
+pub mod repo;