+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.3.3
 
-- Disabled download progress bar due to deadlock in MMRL
 - Improved error handling
-- 
+- Improved MMRL app support
+- Added `repo` subcommand
+- Removed download progress bar due to deadlock in MMRL
+- Other improvements
+- Added ModConf